initial pass at new conductor flow creator
+ break conductor manager view file into four separate files
+ create basic form view for creation of new flows
