+++ purchasing/static/less/_styles.less
@@ -121,10 +121,6 @@ hr.small {
   padding-right: 15px
 }
 
-.dropdown {
-  padding-right: 15px;
-}
-
 .navbar-fixed-top, .navbar-fixed-bottom {
     position: absolute;
 }
