+++ purchasing/static/less/conductor/_flows.less
@@ -1,0 +1,23 @@
+.js-multiform-stage-remove,
+.select2-exterior-container {
+  display: inline-block;
+}
+
+.select2-exterior-container {
+  width: 95%;
+  padding-bottom: 15px;
+
+  &:last-child {
+    padding-bottom: 0;
+  }
+}
+
+
+.select2-container .select2-selection--single,
+.select2-container--default .select2-selection--single .select2-selection__arrow {
+  height: 36px;
+}
+.js-multiform-stage-remove,
+.select2-container--default .select2-selection--single .select2-selection__rendered {
+  line-height: 36px;
+}
