+++ purchasing/static/less/conductor_main.less
@@ -13,6 +13,7 @@
 // Import custom Conductor styles
 @import "conductor/_styles.less";
 @import "conductor/_list_view.less";
+@import "conductor/_flows.less";
 @import "conductor/_detail_view.less";
 @import "conductor/_events.less";
 @import "conductor/_metrics_charts.less";
