+++ purchasing/static/less/opportunities/_styles.less
@@ -176,3 +176,7 @@ h4 {
 .table-beacon a {
   font-weight: 600;
 }
+
+.dropdown-toggle {
+  cursor: pointer;
+}
